[refactor] [#1] sidebarContentTheme 폴더 구조 변경
- sidebarContentTheme 컴포넌트의 폴더 구조를 변경하였다.
- 기존대로 하면 깊이가 깊어질 것 같아서 변경
